requirements.txt
aiogram==2.25.1
aiohttp
transformers
torch
sentencepiece
loguru
newspaper3k
python-decouple
sacremoses